Move label logic out into module, add function for labels by group
